docs(strava-hide-from-feed): enable full readme sync to Greasy Fork
Greasy Fork supports the additional info to be automatically pulled from
an external source, and this README.md at GitHub seems to be a natural
choice.

Greasy Fork only takes the content of the readme though, it doesn't
resolve the images located near the readme. So, fix that by sourcing
the last version of the image on GitHub.